Connected the fragments and data service via the activity. This is done by using LiveData
