+++ app/src/main/res/layout/fragment_contacts.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/linearLayout"
+    android:id="@+id/nav_contact_fragment"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
